Refactoring
* Report.values() returns a list of dicts
* Support for S02 using inheritance
* Refactoring: Moving abstract classes to report/base.py
